Means of the classification accuracy on HAR-dataset
Means of the classification accuracy of an Extra Trees classifier based on the feature selection of QuickSelection, Concrete
Auto-encoder and Autoencoder-based Feature Selection (trained with epoch=10) on the HAR- dataset. The features dropped
are based on the K-value which indicates the amount of features which should be selected in both feature selection algorithms.
The presented results include nine different scenarios, each representing increasing levels of data imbalance (indicated by the
Gini-coefficient, where higher values correspond to greater imbalancedness). The results shows the same dynamics as with the MNIST dataset.