Means of the classification accuracy of an Extra Trees classifier on the HAR-data
Means of the classification accuracy of an Extra Trees classifier based on the feature selection of QuickSelection, Concrete
Auto-encoder and Autoencoder-based Feature Selection (trained with epoch=10) on the HAR-data. The Gini-coefficient
indicates the level of imbalancedness. A Gini-coefficient of 0 signifies perfect balance, where all classes have an equal number of
cases, while a coefficient of 1 indicates maximum imbalance, with all cases belonging to a single class. Four different scenario’s
are shown in which the amount of features which are dropped increases.